mark plugin: move MarkGenerator to model
