+++ _pytest/mark/model.py
@@ -112,3 +112,35 @@ class MarkInfo(object):
         """ yield MarkInfo objects each relating to a marking-call. """
         for args, kwargs in self._arglist:
             yield MarkInfo(self.name, args, kwargs)
+
+
+class MarkGenerator:
+    """ Factory for :class:`MarkDecorator` objects - exposed as
+    a ``pytest.mark`` singleton instance.  Example::
+
+         import py
+         @pytest.mark.slowtest
+         def test_function():
+            pass
+
+    will set a 'slowtest' :class:`MarkInfo` object
+    on the ``test_function`` object. """
+
+    def __init__(self):
+        self.__known_markers = set()
+
+    def __getattr__(self, name):
+        if name[0] == "_":
+            raise AttributeError("Marker name must NOT start with underscore")
+        if hasattr(self, '_config'):
+            self._check(name)
+        return MarkDecorator(name)
+
+    def _check(self, name):
+        if name in self.__known_markers:
+            return
+        from . import _parsed_markers
+        self.__known_markers.update(_parsed_markers(self._config))
+
+        if name not in self.__known_markers:
+            raise AttributeError("%r not a registered marker" % (name, ))
